Moves upgrade-series machine concerns into a new file.
